+++ source/directives/uiTreeNodes.js
@@ -52,9 +52,10 @@
               }
             }, true);
 
-            attrs.$observe('horizontal', function (val) {
-              scope.horizontal = ((typeof val) != 'undefined');
-            });
+            //TODO(jcarter): Remove until horizontal tree is added.
+            // attrs.$observe('horizontal', function (val) {
+            //   scope.horizontal = ((typeof val) != 'undefined');
+            // });
 
           }
         };