+++ source/services/helper.js
@@ -59,20 +59,28 @@
           },
 
           /**
-           * get the event object for touches
-           * @param  {[type]} e [description]
-           * @return {[type]}   [description]
+           * Get the event object for touches.
+           * 
+           * @param  {MouseEvent|TouchEvent} e MouseEvent or TouchEvent that kicked off dragX method.
+           * @return {MouseEvent|TouchEvent} Object returned as original event object.
            */
           eventObj: function (e) {
             var obj = e;
             if (e.targetTouches !== undefined) {
+              //Set obj equal to the first Touch object in the TouchList.
               obj = e.targetTouches.item(0);
+            //Logic to set obj to original TouchEvent.
             } else if (e.originalEvent !== undefined && e.originalEvent.targetTouches !== undefined) {
               obj = e.originalEvent.targetTouches.item(0);
             }
             return obj;
           },
 
+          /**
+           * Generate object used to store data about node being moved.
+           * 
+           * {angular.$scope} node Scope of the node that is being moved.
+           */
           dragInfo: function (node) {
             return {
               source: node,
@@ -83,15 +91,28 @@
                 nodesScope: node.$parentNodesScope
               },
               index: node.index(),
+
+              //slice(0) just duplicates an array.
               siblings: node.siblings().slice(0),
               parent: node.$parentNodesScope,
 
-              // Move the node to a new position
+              //Added as a fix to problem described below:
+              //Source tree is a no drop, drag item over tree that can accept drop, parent is set, then move and drop
+              //item into source tree (meaning outOfBounds is false), because source is no drop and last parent set was
+              //other tree, item is inserted into other tree.
+              resetParent: function() {
+                this.parent = node.$parentNodesScope;
+              },
+
+              //Move the node to a new position, determining where the node will be inserted to when dropped happens here.
               moveTo: function (parent, siblings, index) {
                 this.parent = parent;
+
+                //Duplicate siblings array.
                 this.siblings = siblings.slice(0);
 
-                // If source node is in the target nodes
+                //If source node is in the target nodes
+                //TODO(jcarter): does this take care of re-ordering a tree?
                 var i = this.siblings.indexOf(this.source);
                 if (i > -1) {
                   this.siblings.splice(i, 1);
@@ -104,10 +125,12 @@
                 this.index = index;
               },
 
+              //Get parent nodes nodeScope.
               parentNode: function () {
                 return this.parent.$nodeScope;
               },
 
+              //Get previous sibling node.
               prev: function () {
                 if (this.index > 0) {
                   return this.siblings[this.index - 1];
@@ -116,6 +139,7 @@
                 return null;
               },
 
+              //Get next sibling node.
               next: function () {
                 if (this.index < this.siblings.length - 1) {
                   return this.siblings[this.index + 1];
@@ -124,23 +148,28 @@
                 return null;
               },
 
+              //Return what cloneEnabled is set to on uiTree.
               isClone: function () {
                 return this.source.$treeScope.cloneEnabled === true;
               },
 
+              //Returns a copy of node passed in.
               clonedNode: function (node) {
                 return angular.copy(node);
               },
 
+              //Returns true if parent or index have changed (move happened within any uiTree).
               isDirty: function () {
                 return this.source.$parentNodesScope != this.parent ||
                   this.source.index() != this.index;
               },
 
+              //Return whether node has a new parent (set on moveTo method).
               isForeign: function () {
                 return this.source.$treeScope !== this.parent.$treeScope;
               },
 
+              //Sets arguments passed to user callbacks.
               eventArgs: function (elements, pos) {
                 return {
                   source: this.sourceInfo,
@@ -153,24 +182,27 @@
                 };
               },
 
+              //Method that actually manipulates the node being moved.
               apply: function () {
 
                 var nodeData = this.source.$modelValue;
 
-                // nodrop enabled on tree or parent
+                //Nodrop enabled on tree or parent
+                //Added Object.keys check because parent may be empty object.
                 if (this.parent.nodropEnabled || this.parent.$treeScope.nodropEnabled) {
                   return;
                 }
 
-                // node was dropped in the same place - do nothing
+                //Node was dropped in the same place - do nothing.
                 if (!this.isDirty()) {
                   return;
                 }
 
-                // cloneEnabled and cross-tree so copy and do not remove from source
+                //CloneEnabled and cross-tree so copy and do not remove from source.
                 if (this.isClone() && this.isForeign()) {
                   this.parent.insertNode(this.index, this.sourceInfo.cloneModel);
-                } else { // Any other case, remove and reinsert
+                //Any other case, remove and reinsert.
+                } else {
                   this.source.remove();
                   this.parent.insertNode(this.index, nodeData);
                 }
@@ -244,9 +276,10 @@
            */
           positionStarted: function (e, target) {
             var pos = {},
-              pageX = e.pageX,
-              pageY = e.pageY;
+            pageX = e.pageX,
+            pageY = e.pageY;
 
+            //Check to set correct data for TouchEvents
             if (e.originalEvent && e.originalEvent.touches && (e.originalEvent.touches.length > 0)) {
               pageX = e.originalEvent.touches[0].pageX;
               pageY = e.originalEvent.touches[0].pageY;
@@ -261,44 +294,54 @@
           },
 
           positionMoved: function (e, pos, firstMoving) {
+
             var pageX = e.pageX,
-              pageY = e.pageY,
-              newAx;
+            pageY = e.pageY,
+            newAx;
+
+            //If there are multiple touch points, choose one to use as X and Y.
             if (e.originalEvent && e.originalEvent.touches && (e.originalEvent.touches.length > 0)) {
               pageX = e.originalEvent.touches[0].pageX;
               pageY = e.originalEvent.touches[0].pageY;
             }
-            // mouse position last events
+
+            //Mouse position last event.
+            //TODO(jcarter): What is this being used for?.
             pos.lastX = pos.nowX;
             pos.lastY = pos.nowY;
 
-            // mouse position this events
+            //Mouse position this event.
             pos.nowX = pageX;
             pos.nowY = pageY;
 
-            // distance mouse moved between events
+            //Distance mouse moved between events.
+            //TODO(jcarter): What is this being used for?.            
             pos.distX = pos.nowX - pos.lastX;
             pos.distY = pos.nowY - pos.lastY;
 
-            // direction mouse was moving
+            //Direction mouse was moving.
+            //TODO(jcarter): What is this being used for?.            
             pos.lastDirX = pos.dirX;
             pos.lastDirY = pos.dirY;
 
-            // direction mouse is now moving (on both axis)
+            //Direction mouse is now moving (on both axis).
+            //TODO(jcarter): What is this being used for?.            
             pos.dirX = pos.distX === 0 ? 0 : pos.distX > 0 ? 1 : -1;
             pos.dirY = pos.distY === 0 ? 0 : pos.distY > 0 ? 1 : -1;
 
-            // axis mouse is now moving on
+            //Axis mouse is now moving on.
+            //TODO(jcarter): What is this being used for?.            
             newAx = Math.abs(pos.distX) > Math.abs(pos.distY) ? 1 : 0;
 
-            // do nothing on first move
+            //Do nothing on first move.
             if (firstMoving) {
               pos.dirAx = newAx;
               pos.moving = true;
               return;
             }
 
-            // calc distance moved on this axis (and direction)
+            //Calc distance moved on this axis (and direction).
+            //TODO(jcarter): What is this being used for?.            
             if (pos.dirAx !== newAx) {
               pos.distAxX = 0;
               pos.distAxY = 0;
@@ -307,13 +350,11 @@
               if (pos.dirX !== 0 && pos.dirX !== pos.lastDirX) {
                 pos.distAxX = 0;
               }
-
               pos.distAxY += Math.abs(pos.distY);
               if (pos.dirY !== 0 && pos.dirY !== pos.lastDirY) {
                 pos.distAxY = 0;
               }
             }
-
             pos.dirAx = newAx;
           },
 
@@ -344,25 +385,26 @@
     ]);
 
   // TODO: optimize this loop
+  //(Jcarter): Suggest adding a parent element property on uiTree, then all these bubble to <html> can trigger to stop when they reach the parent.
   function findFirstParentElementWithAttribute(attributeName, childObj) {
-    // undefined if the mouse leaves the browser window
+    //Undefined if the mouse leaves the browser window
     if (childObj === undefined) {
       return null;
     }
     var testObj = childObj.parentNode,
-      count = 1,
-      // check for setAttribute due to exception thrown by Firefox when a node is dragged outside the browser window
-      res = (typeof testObj.setAttribute === 'function' && testObj.hasAttribute(attributeName)) ? testObj : null;
+    count = 1,
+    //Check for setAttribute due to exception thrown by Firefox when a node is dragged outside the browser window
+    res = (typeof testObj.setAttribute === 'function' && testObj.hasAttribute(attributeName)) ? testObj : null;
     while (testObj && typeof testObj.setAttribute === 'function' && !testObj.hasAttribute(attributeName)) {
       testObj = testObj.parentNode;
       res = testObj;
+      //Stop once we reach top of page.
       if (testObj === document.documentElement) {
         res = null;
         break;
       }
       count++;
     }
-
     return res;
   }