+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 2.19.0
+
+* Added check that skips dragDelay logic if no drag delay was specified by user.  This removed unavoidable 100-200ms delay at the start of every drag.
+* Removed all horizontal tree logic that doesn't appear to be supported.
+
 # 2.17.0
 
 * Updated Angular/angularfire references in bower.json